auth-crud-message-div

+++ baatcheet/base/views.py
@@ -1,7 +1,63 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
 from django.db.models import Q
-from .models import Room, Topic, User
+from .models import Room, Topic, Message
+from django.contrib.auth.models import User
 from .forms import RoomForm
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+
+
+def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Something is off, No Records Found!')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Some detail is incorrect, retry!')
+
+    loginPage_data = {'page':page}
+    return render(request, 'base/login_register.html', loginPage_data )
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect ('home')
+        else:
+            messages.error(request, 'Error Occured in Registration!')
+
+
+
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -9,8 +65,8 @@ def home(request):
         Q(topic__name__icontains=q) | 
         Q(name__icontains=q) |
         Q(description__icontains=q)
-        
         )
+        
     topic = Topic.objects.all()
     room_count = rooms.count()
     user = User.objects.filter(
@@ -22,9 +78,22 @@ def home(request):
 
 def room(request, id):
     room = Room.objects.get(id=id)
-    room_data = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body'),
+        )
+        room.participants.add(request.user)
+        return redirect ('room', id=room.id)
+
+    room_data = {'room': room, 'room_messages': room_messages, 'participants':participants}
     return render(request, 'base/room.html', room_data)
-    
+
+@login_required(login_url="login")    
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -38,9 +107,13 @@ def createRoom(request):
     
     return render(request, 'base/room_form.html', create_data)
 
+@login_required(login_url="login") 
 def updateRoom(request, id):
     room = Room.objects.get(id=id)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Not Allowed!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -51,9 +124,13 @@ def updateRoom(request, id):
     update_data = {'form': form}
     return render(request, 'base/room_form.html', update_data)
 
-
+@login_required(login_url="login") 
 def deleteRoom(request, id):
     room = Room.objects.get(id=id)
+
+    if request.user != room.host:
+        return HttpResponse('Not Allowed!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
